extract network connection info component
